data_processor/requirements.txt
numpy==1.18.2
tensorflow==2.1.0
tensorflow-gpu==2.1.0
pytorch-transformers==1.2.0
h5py==2.10.0
pyyaml==3.13
transformers
faiss
tensorflow==2.2.0
tensorflow-gpu==2.2.0
